Fix build warnings for AGP 9.0 and Kotlin K2 compiler
Build system:

- Replace deprecated BaseExtension/LibraryExtension with CommonExtension API

- Remove redundant kotlin-android plugin from all modules (built-in with AGP 9.0)

- Remove deprecated gradle.properties flags (newDsl, builtInKotlin, uniquePackageNames, r8.optimizedResourceShrinking)

Kotlin source:

- Remove unnecessary !! on null-checked receivers

- Add @OptIn(ExperimentalUnsignedTypes) for UByteArray usage

- Remove redundant .toInt() calls on integer expressions

- Replace deprecated Material Icons with AutoMirrored variants

- Remove redundant else from exhaustive when expressions

- Fix unnecessary safe calls where K2 smart-casts apply

- Add scoped @Suppress for unchecked casts and deferred deprecation migrations

+++ sync-core/src/main/java/eu/darken/octi/sync/core/StalenessUtil.kt
@@ -37,7 +37,7 @@ object StalenessUtil {
             }
 
             daysSinceLastSync < 365 -> {
-                val months = (daysSinceLastSync / 30).toInt()
+                val months = daysSinceLastSync / 30
                 context.resources.getQuantityString(
                     R.plurals.sync_stale_period_months,
                     months,
@@ -46,7 +46,7 @@ object StalenessUtil {
             }
 
             else -> {
-                val years = (daysSinceLastSync / 365).toInt()
+                val years = daysSinceLastSync / 365
                 context.resources.getQuantityString(
                     R.plurals.sync_stale_period_years,
                     years,
